:recycle: Move Travel Desk invoce and PNR download to its own component.
NOTE: I opted not to clean up the code at this point, just isolate it.
